Fixed Multiple Bugs
Line 16
Payment Method name was not being added to invoice because of missing parameter on addInvoicePayment()
This is why a workthrough has been made to add fees as zero, to complete the proper parameter for addInvoicePayment() so that the invoice gets marked as paid.


Line 18
I have changed !=325 with ==01 because there exist many more failure message codes such as code 269 which means , multiple withdrawal within short time is not allowed.
So in this case the control should go to 'else' part of the loop because it's failure, but previously it was also going inside the 'if' condition.


Line 20
Paytm Gateway has bug from their side and they need unique transaction id on every request, suppose if a user tries to pay the invoice and then cancels the payment for some reason and then tries again then the paytm gateway will say it's duplicate order id and will reject the payment. So to fix this we added 3 extra characters ( A to Z ) at the end of ORDERID or ORDER_ID when making the initial request from the Pay Button (located in different file of paytm gateway module).
For example if the invoice no is 1025 the new order id would be 1025XYZ , so now if the user cancels the payment and goes back to invoice, a new order id will get generated, so user can make the payment again and now after the transaction is completed we need to remove the extra 3 characters from the $response['ORDERID']  and then apply the payment. 

Note : We cannot update the $response['ORDERID']  because we need to do a checksum match. So we are just changing the $txnid because it will be used to search for invoice and apply payment to it
	


Line 82 
Modified to display the error message that the transaction is successful 

Line 85 ( else block modified )
Modified to display the error message that the transaction failed and redirect the user back to invoice for trying again.