Update paytm.php
Line 21
Paytm Gateway has bug from their side and they need unique transaction id on every request, suppose if a user tries to pay the invoice and then cancels the payment for some reason and then tries again then the paytm gateway will say it's duplicate order id and will reject the payment. so to fix this we have added extra characters ( A to Z ) at the end of ORDERID or ORDER_ID when making the initial request from the PayButton
For example if the invoice no is 1025 the new order id would be 1025XYZ  so  now if the user cancels the payment and goes back to invoice, a new order id will get generated for new transaction.